Possible fix for highlander
TS Fix

Law Maker Incendiary Fix

Possible Lawgiver burstfire fix

Lawgiver Changes:
Fix Rapid Fire
Add Armor Piercing
Add Pellets

Lawgiver:
Nerf incendiary bullets

Blow up the users weapon arm if his dna doesnt match
Added a proper soundeffect

Added Changelog.
Closes #71, Closes #327

Make the lawgiver emagable and change stun to a beam

Added crowdcontrol mode to the lawgiver

Final Tweaks to the lawgiver

Updated .gitignore to inclode the .atom-build.json

Fixed indentation
Fixed blowing up the wrong hand if the user switch hands after the killswitch has been triggered

+++ code/game/antagonist/station/highlander.dm
@@ -35,6 +35,8 @@ var/datum/antagonist/highlander/highlanders
 	for (var/obj/item/I in player)
 		if (istype(I, /obj/item/weapon/implant))
 			continue
+		if (istype(I, /obj/item/organ))
+			continue
 		qdel(I)
 
 	player.equip_to_slot_or_del(new /obj/item/clothing/under/kilt(player), slot_w_uniform)
